moving shared server code into a utils file
